/mnt  share folder on VirtualBox
I ran through an issue when I was trying to have a shared folder from my main computer which is a windows OS through my Vbox Ubuntu. 
This problem mostly occurs with the Ubuntu System. Kali linux is straight forward in regards to this problem 
follow the pdf steps and I also pasted the commands as well